make shared code.
